requirements.txt
google-generativeai==0.8.5
pandas==2.2.3
openpyxl==3.1.5
xlrd== 2.0.1